+++ src/main/java/com/epam/aidial/metric/model/DatasetInfo.java
@@ -1,11 +1,11 @@
-package com.epam.aidial.metric.web.dto;
+package com.epam.aidial.metric.model;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class DatasetInfoDto {
+public class DatasetInfo {
     private String name;
     private Long maxTimeRangeMs;
 }
