+++ src/main/java/com/epam/aidial/metric/web/controller/MetricController.java
@@ -13,7 +13,7 @@ import com.epam.aidial.metric.service.MetricService;
 import com.epam.aidial.metric.web.dto.ColumnDeclarationDto;
 import com.epam.aidial.metric.web.dto.DataDto;
 import com.epam.aidial.metric.web.dto.DataQuery;
-import com.epam.aidial.metric.web.dto.DatasetInfoDto;
+import com.epam.aidial.metric.model.DatasetInfo;
 import com.epam.aidial.metric.web.dto.JsonDataQuery;
 import com.epam.aidial.metric.web.dto.SqlDataQuery;
 import com.epam.aidial.metric.web.dto.TableSchemaDto;
@@ -54,7 +54,7 @@ public class MetricController {
 
     @GetMapping(path = "/datasets",
             produces = MediaType.APPLICATION_JSON_VALUE)
-    public List<DatasetInfoDto> getDatasets() {
+    public List<DatasetInfo> getDatasets() {
         return metricService.getDatasets();
     }
